Update BaseWeaponBP.uasset
For some reason that is beyond me, when I transferred this back to the Mac there was an issue with the BaseWeaponBP that derives from the Weapon C++ class.

While this isn't an issue right now, it could quickly become one once there are a lot of blueprint instances of C++ classes on a map.