Reloading work continues
Added a new action mapping for a reload button, created a new reload montage and updated the anim instance state machine to include it, moved the EAmmoType enum out of Shooter character and into its own header file, worked on a few todo's for reloading.

Still lots to do, but it's getting there I think.