Finished base reloading sequence
Added in a slew of sounds to be used, applied a few of them to reloading, picking up, and equipping a weapon. Reloading sounds occur in the reload montage and the picking up/equipping sounds are setup and called programmatically and assigned in blueprints. Created an animation instance blueprint for the SMG to hand the clip movement and finalize attaching the scene component. Fixed an issue that would cause the game to crash when attempting a reload by actually initializing the scene component in the constructor of the shooter character.